FIX: code "block" detection before showing autocomplete
# What

When a user is composing a message and types a `@`, a `#`, or a `:`, we'll show the proper autocomplete (namely, the users/groups, the categories/tags/channels, and the emojis).

But if the caret position is within a code "block", we should not show the autocomplete because these characters are literals and aren't interpreted.

The Discourse flavour of Markdown supports the following code "blocks":

- Indented code blocks (at least 4 indents more than the preceding block level)
- Code fences (using ``` or ~~~)
- Inline code spans (enclosed in `  `)
- BBCode code "blocks" (enclosed in [code] and [/code]), which can be either inline or block

To support these cases, we used a couple of regular expressions to check whether the caret position was inside a closed or open code block.

This worked OK-ish for simple and well-formed Markdown but failed (somewhat easily) when you start taking into account nested lists and blockquotes.

# How

The correct solution to this problem is to use the same Markdown parser the composer uses.

There are a couple of challenges:

1. Async code: the parsing code is all asynchronous, but the autocomplete library isn't.
2. Performance: we don't want to be parsing the whole document on every single keystroke.

Solving 1. was somewhat straightforward. I basically brute-forced 😅 my way into adding `async / await` in the functions that were using/calling the `inCodeBlock` utility function.

Solving 2. was much easier than I anticipated. The `triggerRule` option is only called when the caret is after the character that has been defined for the autocomplete. I had to remove a call to `inCodeBlock` in a `onKeyUp` event handler, though. If I'm being honest, I don't completely understand how that event handler works and why it's useful (or not).

Now, instead of running some magic regexps, I first try to find a character that isn't in the text so that I can use it as a marker for the position of the caret. I use one of the "private use area" of the Unicode to find such a unique character. If I can't find one, then I bail out.

If I can find a marker, I add it in place of the caret position (the `pos` argument), and then I run a small regular expression that will split the markdown string by empty lines. Empty lines are a very easy way to safely identify blocks in a Markdown document. Then I search for the block that has the marker so that I can parse just that one block instead of the whole document. The parser gives me back a list of tokens (in a tree form, using `children` for recursion), and I use the `findMarker` method to recursively check which token holds the marker. To prevent a stack overflow, I limit the depth of the search to 50 levels.

Once the token that holds the marker is found, I check that its type is either `code_inline`, `code_block`, or `fence`.

# What I Tried

- I tried adding more cases to the regular expressions but kept finding edge cases.
- I tried building a line-by-line parser that would check for Markdown block structures (list items, blockquotes) and code blocks (indented, fences, BBCode), but that was clumsy at best.
- I tried building a trimmed-down version of a Markdown parser that would concentrate solely on the "block structure" of the document. It was easy at first, but when I read the [CommonMark Spec](https://spec.commonmark.org/current/) and the [markdown-it](https://github.com/markdown-it/markdown-it) source code, I stopped because I was about to re-write it.

# Anything Left?

- I probably broke some autocomplete tests with my changes, but I can't seem to be able to run the `qunit` tests locally - all my browsers (Safari, Chrome, and Firefox) crash after ~2000 tests 💥.
- Do the same for #chat? The same autocomplete library is being used, but there are no `triggerRule` associated. I didn't change it in this PR because #chat uses a slightly different version of the Markdown ruleset.